web3Tutor/models.py
from fastapi import FastAPI, Depends, HTTPException, Request, Form, WebSocket, WebSocketDisconnect
from fastapi.responses import RedirectResponse, HTMLResponse, JSONResponse
from fastapi.templating import Jinja2Templates
from fastapi.staticfiles import StaticFiles
from fastapi.middleware.cors import CORSMiddleware
from sqlalchemy import create_engine, Column, Integer, String, ForeignKey, Text, Boolean, Float
from sqlalchemy.orm import sessionmaker, relationship, Session, declarative_base
from pydantic import BaseModel
from passlib.context import CryptContext
import json
import logging
from datetime import datetime
from typing import List, Optional, Dict
from dotenv import load_dotenv
import os
import openai
import schemas, crud

# Existing content from bbb.py follows...