web3Tutor/requirements.txt

fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pydantic==2.6.4
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
jinja2==3.1.3
aiofiles==23.2.1
python-dotenv==1.0.1
web3==6.15.1
openai==1.12.0
